refactor: 무한 스크롤 로직 리팩토링
